Separate js files into util and scripts
