+++ src/js/scripts/simple-video.js
