+++ src/js/util/focus.js
