+++ src/js/util/slider.js
