+++ src/js/util/stringUtil.js
