+++ src/views/cafe.ejs
@@ -96,6 +96,6 @@
       <%- include ("./partials/map.ejs") %> <%- include
       ("./partials/footer.ejs") %>
     </footer>
-    <script src="../js/index.js"></script>
+    <script src="../js/scripts/index.js"></script>
   </body>
 </html>
