+++ src/views/contact.ejs
@@ -66,7 +66,7 @@
       </form>
     </main>
     <footer class="footer"><%- include("./partials/footer") %></footer>
-    <script src="../js/index.js"></script>
-    <script src="../js/form.js"></script>
+    <script src="../js/scripts/index.js"></script>
+    <script src="../js/scripts/form.js"></script>
   </body>
 </html>
