+++ src/views/impressum.ejs
@@ -138,6 +138,6 @@
       </div>
     </main>
     <footer class="footer"><%- include("./partials/footer") %></footer>
-    <script src="../js/index.js"></script>
+    <script src="../js/scripts/index.js"></script>
   </body>
 </html>
