+++ src/views/index.ejs
@@ -134,7 +134,7 @@
       <%- include ("./partials/map.ejs") %> <%- include
       ("./partials/footer.ejs") %>
     </footer>
-    <script src="../js/index.js"></script>
-    <script src="../js/product-slider.js"></script>
+    <script src="../js/scripts/index.js"></script>
+    <script src="../js/scripts/product-slider.js"></script>
   </body>
 </html>
