+++ src/views/origins.ejs
@@ -85,7 +85,7 @@
       </div>
     </main>
     <footer class="footer"><%- include("./partials/footer") %></footer>
-    <script src="../js/index.js"></script>
-    <script src="../js/simple-video.js"></script>
+    <script src="../js/scripts/index.js"></script>
+    <script src="../js/scripts/simple-video.js"></script>
   </body>
 </html>
