+++ src/views/product.ejs
@@ -58,7 +58,7 @@
       </section>
     </main>
     <footer class="footer"><%- include ("./partials/footer.ejs") %></footer>
-    <script src="../js/index.js"></script>
-    <script src="../js/product-details.js"></script>
+    <script src="../js/scripts/index.js"></script>
+    <script src="../js/scripts/product-details.js"></script>
   </body>
 </html>
