+++ src/views/shop.ejs
@@ -16,7 +16,7 @@
       <%- include ("./partials/map.ejs") %> <%- include
       ("./partials/footer.ejs") %>
     </footer>
-    <script src="../js/shop.js"></script>
-    <script src="../js/index.js"></script>
+    <script src="../js/scripts/shop.js"></script>
+    <script src="../js/scripts/index.js"></script>
   </body>
 </html>
